graphical_output: fix strange behavior with direct calls to show_far_fields
when plotting the 'bottom' part, the show_opts variable does not have the
value you'd expect. I guess this has something to do with the recursive
nature of show_far_fields. This alternative way of handling the show_opts
variable is more synthetic and safer against this kind of problems.