avoid MatplotlibDeprecationWarning associated to the use of axes()
MatplotlibDeprecationWarning: Adding an axes using the same arguments
as a previous axes currently reuses the earlier instance. In a future
version, a new instance will always be created and returned. Meanwhile,
this warning can be suppressed, and the future behavior ensured, by
passing a unique label to each axes instance.