converge_l_max: fix bug when target_value is specified
the 'problem' here was that when a target_value is provided, one can
already check l_max=1 against it, instead of waiting for the l_max=2
value to perform the first relative difference calculation. Failing
to recognize this was leading to premature termination of the routine
(l_max=1 was often selected) despite tolerance_step=2.