show_far_field: reintroduce log_scale flag for back-compatibility
make it compatible with passing a 'norm' show option.

it may make sense that the same vmin and vmax values of the 2D maps are
applied to the 1D curves as well, but probably this would generally lead
to a less satisfying result, so I've added the command but commented it out.
This means that at the moment the user cannot change the default y limits
for the 1D curves.